+++ src/Interfaces/card/Card.ts
@@ -1,0 +1,6 @@
+export default interface Card {
+    last4?: number;
+    brand?: string;
+    expMonth?: number;
+    expYear?: number;
+}