+++ src/Models/card/Card.ts
@@ -1,0 +1,7 @@
+import CardInterface from '../../Interfaces/card/Card';
+
+export default class Card {
+    constructor(parameters: CardInterface) {
+        Object.assign(this, parameters);
+    }
+}